fix: changed grade in Activity model to FloatField

+++ users/models.py
@@ -13,6 +13,6 @@ class Course(models.Model):
 
 class Activity(models.Model):
     repo = models.CharField(max_length=255)
-    grade = models.IntegerField(null=True, blank=True, default=None)
+    grade = models.FloatField(null=True, blank=True, default=None)
 
     user_id = models.ForeignKey(User, on_delete=models.CASCADE)
